+++ pioreactor/cli/pios.py
@@ -10,6 +10,7 @@ general API:
 from __future__ import annotations
 
 from concurrent.futures import ThreadPoolExecutor
+from typing import Any
 
 import click
 from msgspec.json import encode as dumps
@@ -20,7 +21,6 @@ from pioreactor.config import config
 from pioreactor.config import get_leader_hostname
 from pioreactor.exc import RsyncError
 from pioreactor.logging import create_logger
-from pioreactor.structs import ArgsOptions
 from pioreactor.mureq import HTTPException
 from pioreactor.pubsub import post_into
 from pioreactor.utils import ClusterJobManager
@@ -67,7 +67,7 @@ if am_I_leader() or is_testing_env():
     confirmation = click.option("-y", is_flag=True, help="Skip asking for confirmation.")
     json_output = click.option("--json", is_flag=True, help="output as json")
 
-    def parse_click_arguments(input_list: list[str]) -> ArgsOptions:
+    def parse_click_arguments(input_list: list[str]) -> dict:  # TODO: typed dict
         args: list[str] = []
         opts: dict[str, str | None] = {}
 
@@ -91,7 +91,7 @@ if am_I_leader() or is_testing_env():
 
             i += 1
 
-        return ArgsOptions(args=args, options=opts)
+        return {"args": args, "options": opts}
 
     def universal_identifier_to_all_active_workers(workers: tuple[str, ...]) -> tuple[str, ...]:
         active_workers = get_active_workers_in_inventory()
@@ -281,13 +281,13 @@ if am_I_leader() or is_testing_env():
             def _thread_function(unit: str) -> tuple[bool, dict]:
                 try:
                     r = post_into(
-                        resolve_to_address(unit), "/unit_api/system/update", json=ArgsOptions(options=options)
+                        resolve_to_address(unit), "/unit_api/system/update", json={"options": options}
                     )
                     r.raise_for_status()
                     return True, r.json()
-                except HTTPException:
-                    logger.error(f"Unable to update {target} on {unit} due to server error: {e}.")
-                    return False
+                except HTTPException as e:
+                    logger.error(f"Unable to update on {unit} due to server error: {e}.")
+                    return False, {"unit": unit}
 
             with ThreadPoolExecutor(max_workers=len(units)) as executor:
                 results = executor.map(_thread_function, units)
@@ -349,7 +349,7 @@ if am_I_leader() or is_testing_env():
         def _thread_function(unit: str) -> tuple[bool, dict]:
             try:
                 r = post_into(
-                    resolve_to_address(unit), "/unit_api/system/update/app", json=ArgsOptions(options=options)
+                    resolve_to_address(unit), "/unit_api/system/update/app", json={"options": options}
                 )
                 r.raise_for_status()
                 return True, r.json()
@@ -417,7 +417,7 @@ if am_I_leader() or is_testing_env():
         def _thread_function(unit: str) -> tuple[bool, dict]:
             try:
                 r = post_into(
-                    resolve_to_address(unit), "/unit_api/system/update/ui", json=ArgsOptions(options=options)
+                    resolve_to_address(unit), "/unit_api/system/update/ui", json={"options": options}
                 )
                 r.raise_for_status()
                 return True, r.json()
@@ -463,10 +463,10 @@ if am_I_leader() or is_testing_env():
                 raise click.Abort()
 
         logger = create_logger("install_plugin", unit=get_unit_name(), experiment=UNIVERSAL_EXPERIMENT)
-        commands = ArgsOptions(args=[plugin])
+        commands: dict[str, Any] = {"args": [plugin]}
 
         if source:
-            commands.options["source"] = source
+            commands["options"] = {"source": source}
 
         def _thread_function(unit: str) -> tuple[bool, dict]:
             try:
@@ -508,7 +508,7 @@ if am_I_leader() or is_testing_env():
                 raise click.Abort()
 
         logger = create_logger("uninstall_plugin", unit=get_unit_name(), experiment=UNIVERSAL_EXPERIMENT)
-        commands = ArgsOptions(args=[plugin])
+        commands = {"args": [plugin]}
 
         def _thread_function(unit: str) -> tuple[bool, dict]:
             try: