+++ pioreactor/utils/pwm.py
@@ -195,7 +195,7 @@ class PWM:
         self.duty_cycle = 0.0
 
         if self.is_locked():
-            msg = f"GPIO-{self.pin} is currently locked but a task is overwriting it. Either too many jobs are trying to access this pin, or a job didn't clean up properly. If your confident you can release it, use `pio clear-cache pwm_locks {self.pin} --as-int` on the command line for {self.unit}."
+            msg = f"GPIO-{self.pin} is currently locked but a task is overwriting it. Either too many jobs are trying to access this pin, or a job didn't clean up properly. If your confident you can release it, use `pio cache clear pwm_locks {self.pin} --as-int` on the command line for {self.unit}."
 
             self.logger.error(msg)
             raise PWMError(msg)